Split the v5_v6 DB upgrade to its own file
